Move /internal/cmd back to /cmd
