+++ cmd/validate.go
@@ -11,7 +11,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
-	"github.com/jetstack/paranoia/internal/cmd/options"
+	"github.com/jetstack/paranoia/cmd/options"
 	"github.com/jetstack/paranoia/internal/image"
 	"github.com/jetstack/paranoia/internal/validate"
 )
